+++ examples/pytorch/knowledgegraph/main.py
@@ -8,66 +8,32 @@ import numpy as np
 from deepgnn import TrainMode, setup_default_logging_config
 
 from deepgnn.pytorch.modeling import BaseModel
+from deepgnn.pytorch.common.ray_train import run_ray
 from deepgnn.pytorch.common.dataset import TorchDeepGNNDataset
 from deepgnn.graph_engine import GEEdgeSampler, GraphEngineBackend
 from model import KGEModel  # type: ignore
 from deepgnn import get_logger
 
 
-from typing import Dict
-import os
-import platform
-import numpy as np
-import torch
-import ray
-import ray.train as train
-from ray.train.torch import TorchTrainer
-from ray.air import session
-from ray.air.config import ScalingConfig
-from deepgnn import TrainMode, get_logger
-from deepgnn.graph_engine import create_backend, BackendOptions
-from deepgnn.graph_engine.samplers import GENodeSampler, GEEdgeSampler
-from deepgnn.pytorch.common import get_args
-from deepgnn.pytorch.common.utils import load_checkpoint, save_checkpoint
+def create_model(args: argparse.Namespace):
+    get_logger().info(f"Creating KGEModel with seed:{args.seed}.")
+    # set seed before instantiating the model
 
-
-def train_func(config: Dict):
-    """Training loop for ray trainer."""
-    args = config["args"]
-
-    logger = get_logger()
-    os.makedirs(args.save_path, exist_ok=True)
-
-    train.torch.accelerate(args.fp16)
-    if args.seed:
-        train.torch.enable_reproducibility(seed=args.seed + session.get_world_rank())
 
     model_args = json.loads(args.model_args)
-    model = KGEModel(
+    return KGEModel(
         num_negs=args.num_negs, gpu=args.cuda, embed_dim=args.dim, model_args=model_args
     )
-    # https://docs.ray.io/en/latest/tune/api_docs/trainable.html#function-api-checkpointing
-    model = train.torch.prepare_model(model, move_to_device=args.gpu)
-    if args.mode == TrainMode.TRAIN:
-        model.train()
-    else:
-        model.eval()
 
-    epochs_trained, steps_in_epoch_trained = load_checkpoint(
-        model, logger, args, session.get_world_rank()
-    )
 
-    optimizer = torch.optim.Adam(
-        filter(lambda p: p.requires_grad, model.parameters()),
-        lr=args.learning_rate * session.get_world_size(),
-        weight_decay=0,
-    )
-    optimizer = train.torch.prepare_optimizer(optimizer)
-
-    backend = create_backend(
-        BackendOptions(args), is_leader=(session.get_world_rank() == 0)
-    )
-    dataset = TorchDeepGNNDataset(
+def create_dataset(
+    args: argparse.Namespace,
+    model: BaseModel,
+    rank: int = 0,
+    world_size: int = 1,
+    backend: GraphEngineBackend = None,
+):
+    return TorchDeepGNNDataset(
         sampler_class=GEEdgeSampler,
         backend=backend,
         query_fn=model.query if args.mode == TrainMode.TRAIN else model.query_eval,
@@ -76,59 +42,17 @@ def train_func(config: Dict):
         batch_size=args.batch_size,
         edge_types=np.array([args.edge_type], dtype=np.int32),
         epochs=1,
-        sample_num=args.max_id // session.get_world_size(),
-        num_workers=session.get_world_size(),
+        sample_num=args.max_id // world_size,
+        num_workers=world_size,
     )
-    num_workers = 0
-    dataset = torch.utils.data.DataLoader(
-        dataset=dataset,
-        num_workers=num_workers,
+
+
+def create_optimizer(args: argparse.Namespace, model: BaseModel, world_size: int):
+    return torch.optim.Adam(
+        filter(lambda p: p.requires_grad, model.parameters()),
+        lr=args.learning_rate * world_size,
+        weight_decay=0,
     )
-    for epoch in range(epochs_trained, args.num_epochs):
-        scores = []
-        labels = []
-        losses = []
-        for step, batch in enumerate(dataset):
-            if step < steps_in_epoch_trained:
-                continue
-            loss, score, label = model(batch)
-            optimizer.zero_grad()
-            loss.backward()
-            optimizer.step()
-
-            scores.append(score)
-            labels.append(label)
-            losses.append(loss.item())
-
-        steps_in_epoch_trained = 0
-        if epoch % args.save_ckpt_by_epochs == 0:
-            save_checkpoint(model, logger, epoch, step, args)
-
-        session.report(
-            {
-                "metric": model.compute_metric(scores, labels).item(),
-                "loss": np.mean(losses),
-            },
-        )
-
-
-def run_ray(**kwargs):
-    """Run ray trainer."""
-    ray.init()
-
-    args = get_args(None, kwargs["run_args"] if "run_args" in kwargs else None)
-
-    trainer = TorchTrainer(
-        train_func,
-        train_loop_config={
-            "args": args,
-            **kwargs,
-        },
-        scaling_config=ScalingConfig(
-            num_workers=1, use_gpu=args.gpu, resources_per_worker={"CPU": 2}
-        ),
-    )
-    trainer.fit()
 
 
 def _main():
@@ -138,7 +62,11 @@ def _main():
     # run_dist is the unified entry for pytorch model distributed training/evaluation/inference.
     # User only needs to prepare initializing function for model, dataset, optimizer and args.
     # reference: `deepgnn/pytorch/training/factory.py`
-    run_ray()
+    run_ray(
+        init_model_fn=create_model,
+        init_dataset_fn=create_dataset,
+        init_optimizer_fn=create_optimizer,
+    )
 
 
 if __name__ == "__main__":