+++ examples/pytorch/link_prediction/main.py
@@ -11,43 +11,25 @@ from deepgnn.pytorch.common.utils import (
     get_logger,
     get_python_type,
     get_store_name_and_path,
+    
 )
 from deepgnn.pytorch.encoding import get_feature_encoder
 from deepgnn.pytorch.modeling import BaseModel
+from deepgnn.pytorch.common.ray_train import run_ray
 from deepgnn.graph_engine import TextFileSampler, GraphEngineBackend
 from args import init_args  # type: ignore
 from consts import DEFAULT_VOCAB_CHAR_INDEX  # type: ignore
 from model import LinkPredictionModel  # type: ignore
-from typing import Dict
-import os
-import platform
-import numpy as np
-import torch
-import ray
-import ray.train as train
-from ray.train.torch import TorchTrainer
-from ray.air import session
-from ray.air.config import ScalingConfig
-from deepgnn import TrainMode, get_logger
-from deepgnn.graph_engine import create_backend, BackendOptions
-from deepgnn.graph_engine.samplers import GENodeSampler, GEEdgeSampler
-from deepgnn.pytorch.common import get_args
-from deepgnn.pytorch.common.utils import load_checkpoint, save_checkpoint
 
 
-def train_func(config: Dict):
-    """Training loop for ray trainer."""
-    args = config["args"]
+def create_model(args: argparse.Namespace):
+    get_logger().info(f"Creating LinkPredictionModel with seed:{args.seed}.")
+    # set seed before instantiating the model
 
-    logger = get_logger()
-    os.makedirs(args.save_path, exist_ok=True)
-
-    train.torch.accelerate(args.fp16)
-    if args.seed:
-        train.torch.enable_reproducibility(seed=args.seed + session.get_world_rank())
 
     feature_enc = get_feature_encoder(args)
-    model = LinkPredictionModel(
+
+    return LinkPredictionModel(
         args=args,
         feature_dim=args.feature_dim,
         feature_idx=args.feature_idx,
@@ -55,29 +37,18 @@ def train_func(config: Dict):
         feature_enc=feature_enc[0],  # type: ignore
         vocab_index=feature_enc[1][DEFAULT_VOCAB_CHAR_INDEX],  # type: ignore
     )
-    # https://docs.ray.io/en/latest/tune/api_docs/trainable.html#function-api-checkpointing
-    model = train.torch.prepare_model(model, move_to_device=args.gpu)
-    if args.mode == TrainMode.TRAIN:
-        model.train()
-    else:
-        model.eval()
 
-    epochs_trained, steps_in_epoch_trained = load_checkpoint(
-        model, logger, args, session.get_world_rank()
-    )
 
-    optimizer = torch.optim.AdamW(
-        model.parameters(),
-        lr=args.learning_rate * session.get_world_size(),
-        weight_decay=args.weight_decay,
-    )
-    optimizer = train.torch.prepare_optimizer(optimizer)
-
-    backend = create_backend(
-        BackendOptions(args), is_leader=(session.get_world_rank() == 0)
-    )
+def create_dataset(
+    args: argparse.Namespace,
+    model: BaseModel,
+    rank: int = 0,
+    world_size: int = 1,
+    backend: GraphEngineBackend = None,
+):
     store_name, relative_path = get_store_name_and_path(args.train_file_dir)
-    dataset = TorchDeepGNNDataset(
+
+    return TorchDeepGNNDataset(
         sampler_class=TextFileSampler,
         backend=backend,
         query_fn=model.query,
@@ -89,61 +60,19 @@ def train_func(config: Dict):
         adl_config=args.adl_config,
         shuffle=False,
         drop_last=False,
-        worker_index=session.get_world_rank(),
-        num_workers=session.get_world_size(),
+        worker_index=rank,
+        num_workers=world_size,
         epochs=-1 if (args.max_samples > 0 and args.mode == TrainMode.TRAIN) else 1,
         buffer_size=1024,
     )
-    num_workers = args.data_parallel_num
-    dataset = torch.utils.data.DataLoader(
-        dataset=dataset,
-        num_workers=num_workers,
+
+
+def create_optimizer(args: argparse.Namespace, model: BaseModel, world_size: int):
+    return torch.optim.AdamW(
+        model.parameters(),
+        lr=args.learning_rate * world_size,
+        weight_decay=args.weight_decay,
     )
-    for epoch in range(epochs_trained, args.num_epochs):
-        scores = []
-        labels = []
-        losses = []
-        for step, batch in enumerate(dataset):
-            if step < steps_in_epoch_trained:
-                continue
-            loss, score, label = model(batch)
-            optimizer.zero_grad()
-            loss.backward()
-            optimizer.step()
-
-            scores.append(score)
-            labels.append(label)
-            losses.append(loss.item())
-
-        steps_in_epoch_trained = 0
-        if epoch % args.save_ckpt_by_epochs == 0:
-            save_checkpoint(model, logger, epoch, step, args)
-
-        session.report(
-            {
-                "metric": model.compute_metric(scores, labels).item(),
-                "loss": np.mean(losses),
-            },
-        )
-
-
-def run_ray(**kwargs):
-    """Run ray trainer."""
-    ray.init()
-
-    args = get_args(init_args, kwargs["run_args"] if "run_args" in kwargs else None)
-
-    trainer = TorchTrainer(
-        train_func,
-        train_loop_config={
-            "args": args,
-            **kwargs,
-        },
-        scaling_config=ScalingConfig(
-            num_workers=1, use_gpu=args.gpu, resources_per_worker={"CPU": 2}
-        ),
-    )
-    trainer.fit()
 
 
 def _main():
@@ -153,7 +82,12 @@ def _main():
     # run_dist is the unified entry for pytorch model distributed training/evaluation/inference.
     # User only needs to prepare initializing function for model, dataset, optimizer and args.
     # reference: `deepgnn/pytorch/training/factory.py`
-    run_ray()
+    run_ray(
+        init_model_fn=create_model,
+        init_dataset_fn=create_dataset,
+        init_optimizer_fn=create_optimizer,
+        init_args_fn=init_args,
+    )
 
 
 if __name__ == "__main__":