+++ src/python/deepgnn/migrate/0_1_58.py
@@ -23,30 +23,6 @@ if __name__ == "__main__":
     )
     args = parser.parse_args()
 
-    dirname = os.path.dirname(args.main_path)
-    if dirname == "":
-        dirname = "."
-    with open(
-        os.path.join(os.path.dirname(__file__), "data", "ray_util.py"), "r"
-    ) as file:
-        ray_util = file.read()
-    if args.hvd:
-        # https://docs.ray.io/en/latest/_modules/ray/train/horovod/horovod_trainer.html
-
-        ray_util = ray_util.replace(
-            "from ray.train.torch import TorchTrainer",
-            "import horovod.torch as hvd\nimport ray.train.torch\nfrom ray.train.horovod import HorovodTrainer",
-        )
-        ray_util = ray_util.replace("train.torch.accelerate(args.fp16)", "hvd.init()")
-        ray_util = ray_util.replace(
-            "train.torch.prepare_optimizer(optimizer)",
-            "hvd.DistributedOptimizer(optimizer, named_parameters=model.named_parameters())",
-        )
-        ray_util = ray_util.replace("TorchTrainer", "HorovodTrainer")
-
-    with open(os.path.join(dirname, "ray_util.py"), "w") as file:
-        file.write(ray_util)
-
     with open(args.main_path, "r") as file:
         raw_input = file.read()
 
@@ -58,9 +34,14 @@ if __name__ == "__main__":
         "deepgnn.pytorch.training.util.disable_infini_band",
         "deepgnn.pytorch.util.disable_infini_band",
     )
-    rename.rename_external(
-        tree, "deepgnn.pytorch.training.run_dist", "ray_util.run_ray"
-    )
+    if args.hvd:
+        rename.rename_external(
+            tree, "deepgnn.pytorch.training.run_dist", "deepgnn.pytorch.common.horovod_train.run_ray"
+        )
+    else:
+        rename.rename_external(
+            tree, "deepgnn.pytorch.training.run_dist", "deepgnn.pytorch.common.ray_train.run_ray"
+        )
 
     raw_output = pasta.dump(tree)